+++ index.html
@@ -24,6 +24,11 @@
     <script src="https://unpkg.com/babel-standalone@6/babel.min.js"></script>
 
     <!-- Load our React component. -->
-    <script src="todo.js" type="text/babel"></script>
+    
+    <!-- Uncaught ReferenceError: require is not defined -->
+    <!-- <script src="todo.js" type="text/babel"></script> -->
+
+    <!-- Uncaught SyntaxError: Unexpected token '<' -->
+    <script src="todo.js" type="module"></script>
   </body>
 </html>
